requirements.txt
streamlit
biopython
st-py3dmol
python-docx